1.修复Error序号对应错误问题 2.修复NetEvent::ready成功导致NetEvent::resume仍导致IO被执行的问题 3.修复InternalLogger::getInstance非线程安全的问题 4.优化Runtime实现 5.增加BackTrace类 6.增加TcpServer类

+++ galay/utils/Tree.cc
@@ -65,7 +65,7 @@ TrieTree::remove(const std::string& word)
     TrieNode* cur = m_root;
     TrieNode* del = m_root;
     char delCh = word[0];
-    for (int i = 0; i < word.length() ; ++i)
+    for (size_t i = 0; i < word.length() ; ++i)
     {
         auto childIt = cur->m_childMap.find(word[i]);
         if (childIt == cur->m_childMap.end())
